Added seed script for h2 database

+++ vffd-api/src/main/java/com/lanfear/vffd/ForecastRecord.java
@@ -26,7 +26,7 @@ public class ForecastRecord {
     private String featureCode;
     private String featureName;
     private Integer projectedVolume;
-    private Float adoptionRate;
+    private Double adoptionRate;
     private String lastUpdatedBy;
 
 }
